requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==14.3.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.2.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
datasets==2.19.2
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
docopt==0.6.2
evaluate==0.4.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gradio==3.32.0
gradio_client==1.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
imageio==2.37.0
iopath==0.1.10
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
linkify-it-py==2.0.3
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.16
narwhals==1.33.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.4.2
numpy==1.24.4
opencv-python==4.8.1.78
orjson==3.10.16
packaging==24.2
pandas==1.5.3
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pickleshare==0.7.5
pillow==11.1.0
pipreqs==0.5.0
platformdirs==4.3.7
portalocker==3.1.1
prompt_toolkit==3.0.50
propcache==0.3.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyarrow-hotfix==0.6
pydantic==2.11.2
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorchvideo==0.1.5
pytz==2025.2
PyWavelets==1.8.0
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
responses==0.18.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.15.2
semantic-version==2.10.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.1
sympy==1.13.3
tabulate==0.9.0
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
tokenizers==0.19.1
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.40.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uc-micro-py==1.0.3
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
xxhash==3.5.0
yacs==0.1.8
yarg==0.1.9
yarl==1.19.0